Add stricter headers
